src/converter/requirements.txt
black==22.10.0
moviepy==1.0.3
pika==1.3.1
pip-chill==1.0.1
pymongo==4.3.3
python-decouple==3.6
# certifi==2022.9.24 # Installed as dependency for requests
# charset-normalizer==2.1.1 # Installed as dependency for requests
# click==8.1.3 # Installed as dependency for black
# decorator==4.4.2 # Installed as dependency for moviepy
# dnspython==2.2.1 # Installed as dependency for pymongo
# idna==3.4 # Installed as dependency for requests
# imageio==2.22.4 # Installed as dependency for moviepy
# imageio-ffmpeg==0.4.7 # Installed as dependency for moviepy
# mypy-extensions==0.4.3 # Installed as dependency for black
# numpy==1.23.5 # Installed as dependency for moviepy, imageio
# pathspec==0.10.2 # Installed as dependency for black
# pillow==9.3.0 # Installed as dependency for imageio
# platformdirs==2.5.4 # Installed as dependency for black
# proglog==0.1.10 # Installed as dependency for moviepy
# requests==2.28.1 # Installed as dependency for moviepy
# tomli==2.0.1 # Installed as dependency for black
# tqdm==4.64.1 # Installed as dependency for moviepy, proglog
# urllib3==1.26.12 # Installed as dependency for requests
